#2 Make the code clearer and simpler.
-change turn_on_off and change_mode to one line if statement
-split to cpp and hpp file
-adding UNKNOWN state
