+++ rafal_szyinski/pin_controls.hpp
@@ -1,0 +1,5 @@
+#pragma once
+#include "pins.hpp"
+
+void turn_off_color();
+void disco();
